relocate state infrastructure
