+++ src/features/pokemon-list/infrastructure/react/hooks/useListControls.ts
@@ -1,10 +1,15 @@
-import { useAppDispatch, useAppSelector } from "../../../../../infrastructure/redux/hooks";
+import {
+  useAppDispatch,
+  useAppSelector,
+} from "../../../../../shared/infrastructure/redux/hooks";
 import { toggleSortByHeight } from "../../redux/slices/listControlsSlice";
-import { RootState } from "../../../../../infrastructure/redux/store";
+import { RootState } from "../../../../../shared/infrastructure/redux/store";
 
 export const useListControls = () => {
   const dispatch = useAppDispatch();
-  const sortByHeight = useAppSelector((state: RootState) => state.listControls.sortByHeight);
+  const sortByHeight = useAppSelector(
+    (state: RootState) => state.listControls.sortByHeight
+  );
 
   const handleToggleSortByHeight = () => {
     dispatch(toggleSortByHeight());
