+++ src/pages/Home/__tests__/Home.SortByHeight.test.tsx
@@ -3,7 +3,7 @@ import { expect, it } from "vitest";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import { Provider } from "react-redux";
 import Home from "../Home";
-import { store } from "../../../infrastructure/redux/store";
+import { store } from "../../../shared/infrastructure/redux/store";
 import { clickSortByHeightCheckbox } from "./helpers";
 
 it("sorts pokemon list by height when checkbox is clicked", async () => {
